script we use with geocodio
notes: I didn't write this! but I run it every few months. One thing this does that you won't have to worry about is that it checks to see if we have geocoded an address in the past. I also included all of the libraries associated with the larger script because I wasn't sure which functions were used in this piece of the code specifically.